requirements.txt
bittensor==9.7.0
bittensor-cli==9.7.0
substrate-interface==1.7.11
python-dotenv==1.1.1
bcrypt